--- a/src/libraries/System.Diagnostics.DiagnosticSource/src/System/Diagnostics/Activity.cs
+++ b/src/libraries/System.Diagnostics.DiagnosticSource/src/System/Diagnostics/Activity.cs
@@ -39,8 +39,6 @@ namespace System.Diagnostics
         private static readonly IEnumerable<ActivityLink> s_emptyLinks = new ActivityLink[0];
         private static readonly IEnumerable<ActivityEvent> s_emptyEvents = new ActivityEvent[0];
 #pragma warning restore CA1825
-        internal static readonly Random s_random = new Random();
-
         private static readonly ActivitySource s_defaultSource = new ActivitySource(string.Empty);
 
         private const byte ActivityTraceFlagsIsSet = 0b_1_0000000; // Internal flag to indicate if flags have been set
@@ -1665,15 +1663,13 @@ namespace System.Diagnostics
         internal static unsafe void SetToRandomBytes(Span<byte> outBytes)
         {
             Debug.Assert(outBytes.Length == 16 || outBytes.Length == 8);
-            Random r = Activity.s_random;
+            RandomNumberGenerator r = RandomNumberGenerator.Current;
 
             Unsafe.WriteUnaligned(ref outBytes[0],  r.Next());
-            Unsafe.WriteUnaligned(ref outBytes[4],  r.Next());
 
             if (outBytes.Length >= 16)
             {
                 Unsafe.WriteUnaligned(ref outBytes[8],  r.Next());
-                Unsafe.WriteUnaligned(ref outBytes[12], r.Next());
             }
         }