Fixes https://github.com/dotnet/runtime/issues/46818

Activity class uses `Guid.NewGuid()` to generate the Trace and Span Id's. It is proven on Linux Guid generation is very slow https://github.com/dotnet/runtime/issues/13628. It was noticed in some activity scenarios that Guid generation inside the Activity consuming none-trivial amount of time.
The fix here is to change the way we generate the Ids to use Random class instead of Guid.

# Linux

*Before*

|                Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
| ActivityIdsGeneration | 4.386 us | 0.0330 us | 0.0258 us | 0.0992 |     - |     - |     632 B |

*After*

|                Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| ActivityIdsGeneration | 305.9 ns | 4.71 ns | 4.17 ns | 0.0625 |     - |     - |     392 B |


# Windows

*Before*

|                Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| ActivityIdsGeneration | 544.1 ns | 6.12 ns | 5.72 ns | 0.1001 |     - |     - |     632 B |

*After*

|                Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| ActivityIdsGeneration | 355.5 ns | 5.28 ns | 4.94 ns | 0.0625 |     - |     - |     392 B |