@noahfalk 2021-01-13
Is the code for that benchmark you ran shown somewhere @tarekgh? Is it measuring the total time to create, start, and stop an Activity? If the benchmark is only measuring the time to generate 16 bytes of random data I would hope we can do that in 20-50ns if we are being efficient.