@tarekgh 2021-01-13
@noahfalk the code I used is 

```C#
        [Benchmark] public int ActivityIdsGeneration()
        {
            using Activity activity = new Activity("MyActivity");
            activity.Start();
            ActivityTraceId traceId = activity.TraceId;
            ActivitySpanId spanId = activity.SpanId;

            return 100;
        }
```