@tarekgh 2021-01-14
I have implemented the class to generate a long random numbers. Basically I used the same code/algorithm used in the Random class. I have spent sometime testing it and I am seeing it is good enough (I guess). but if you have any tests you want to try, please let me know and I can try it. Also, I made the objects generating the random numbers be thread static. And here is the perf numbers with the changes:

# Linux

**Before**

|                Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
| ActivityIdsGeneration | 4.222 us | 0.0294 us | 0.0261 us | 0.0992 |     - |     - |     632 B |

**After**

|                Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| ActivityIdsGeneration | 299.1 ns | 2.95 ns | 2.61 ns | 0.0625 |     - |     - |     392 B |


# Windows

**Before**

|                Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| ActivityIdsGeneration | 533.7 ns | 2.91 ns | 2.72 ns | 0.1001 |     - |     - |     632 B |

**After**

|                Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| ActivityIdsGeneration | 330.7 ns | 0.56 ns | 0.44 ns | 0.0625 |     - |     - |     392 B |