@tarekgh 2021-01-18
I have changed the used random number generator to the one pointed earlier by @stephentoub. and here is the current perf numbers with that change:

# Linux

**Before**

|                Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
| ActivityIdsGeneration | 4.067 us | 0.0294 us | 0.0275 us | 0.0992 |     - |     - |     632 B |

**After**

|                Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| ActivityIdsGeneration | 285.6 ns | 3.50 ns | 3.11 ns | 0.0625 |     - |     - |     392 B |



# Windows 

**Before**

|                Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| ActivityIdsGeneration | 523.2 ns | 4.18 ns | 3.70 ns | 0.1001 |     - |     - |     632 B |

**After**

|                Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| ActivityIdsGeneration | 334.6 ns | 1.01 ns | 0.90 ns | 0.0625 |     - |     - |     392 B |